1.0.0: Small and simple layout library for Preact
# preact-layout
Preact Layout elegantly solves some common layout challenges you will undoubtedly
encounter when using Preact, due to the hierarchical nature of the component
tree and the one-way data flow. Preact Layout allows you to think out of the box.

### Simple but powerful API
With just 2 components, the API is very simple to learn, yet powerful.
* [Layout](docs/api/Layout.md) to define layouts
* [Section](docs/api/Section.md) to divide the layout into multiple sections

### Lightweight
preact-layout is microscopically small. The sources are just over 2 Kb
and the minified and gzipped distribution file weighs only 1 kB.

### Well documented
A good library needs plenty of good docs. In the case of preact-layout, the
docs are way bigger than the library itself! We have a
[usage guide](https://download.github.io/preact-layout/docs/getting-started/Basic-usage.html),
[API docs](https://download.github.io/preact-layout/docs/api/),
[examples](https://download.github.io/preact-layout/docs/getting-started/Examples.html) and a
[Preact Layout Kickstart](http://codepen.io/StijnDeWitt/pen/rrzJEA?editors=0010) CodePen to learn from.

### Well tested
Good tests ensure not only that the code works, but that it keeps working! We use [Travis
CI](https://travis-ci.org/Download/preact-layout) to test every push to master so we
catch any bugs before they make it into a release.

### Open Source, Open Community
preact-layout is Open Source software with a
[public code repository](https://github.com/download/preact-layout) and
[public issue tracker](https://github.com/download/preact-layout/issues).